requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
Cerberus==1.2
certifi==2018.8.24
chardet==3.0.4
Django==2.1.1
django-celery==3.2.2
django-cors-headers==2.4.0
django-environ==0.4.5
djangorestframework==3.8.2
idna==2.7
kombu==3.0.37
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
PySocks==1.6.8
pytz==2018.5
requests==2.19.1
serpy==0.3.1
six==1.11.0
twilio==6.16.4
urllib3==1.23
vine==1.1.4
